Allow rate to be less than 1.
Change all the token-related variables to doubles so the rate can be
less than 1.

+++ userspace/engine/token_bucket.cpp
@@ -31,7 +31,7 @@ token_bucket::~token_bucket()
 {
 }
 
-void token_bucket::init(uint32_t rate, uint32_t max_tokens)
+void token_bucket::init(double rate, double max_tokens)
 {
 	m_rate = rate;
 	m_max_tokens = max_tokens;
@@ -48,7 +48,7 @@ bool token_bucket::claim(uint64_t now)
 		now = sinsp_utils::get_current_time_ns();
 	}
 
-	uint64_t tokens_gained = (now - m_last_seen) / (m_rate * 1000000000);
+	double tokens_gained = (now - m_last_seen) / (m_rate * 1000000000);
 	m_last_seen = now;
 
 	m_tokens += tokens_gained;
